[fix]: 체크박스 선택 오류 수정
